+++ http_cache_drift_store/lib/src/store/db_platform/db_platform.dart
@@ -1,3 +1,1 @@
-export 'db_none.dart'
-    if (dart.library.html) 'db_web.dart'
-    if (dart.library.io) 'db_os.dart';
+export 'db_web.dart' if (dart.library.io) 'db_os.dart';