+++ http_cache_sembast_store/CHANGELOG.md
@@ -1,5 +1,7 @@
-## next
-- chore: Min SDK is now 3.6.0.
+## 1.0.1
+- fix: don't misreport older store unknown status code entry as 304.
+- fix: Store race condition when cleaning at start.
+- chore: Min SDK is now 3.8.0.
 
 ## 1.0.0
 - feat: Saves response status code.